pyproject.toml
[tool.poetry]
name = "kata: number to french converter"
version = "0.0.1"
description = "A tool to convert numbers into their French word representations."
authors = ["Zakaria Bouhoun"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
click = "^8.1.7"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
